backend/requirements.txt
Adafruit-Blinka==8.18.1
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-requests==1.13.2
adafruit-circuitpython-typing==1.9.2
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
bandit==1.7.5
billiard==4.1.0
celery==5.3.0b2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2
django-appconf==1.0.5
django-celery-results==2.5.0
django-compressor==4.3.1
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
kombu==5.3.0b3
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
pbr==5.11.1
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pyflakes==3.0.1
pyftdi==0.54.0
Pygments==2.15.1
pyserial==3.5
pytz==2023.3
pyusb==1.2.1
PyYAML==6.0
rcssmin==1.1.1
redis==4.5.4
rich==13.3.5
rjsmin==1.2.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
stevedore==5.0.0
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
